Fix: Minor fixes across all DTOs

+++ src/main/java/com/poketeambuilder/dtos/front/ability/AbilityReadDto.java
@@ -1,5 +1,5 @@
 package com.poketeambuilder.dtos.front.ability;
 
-public record AbilityReadDto(long id, String name, String effect, String description) {
+public record AbilityReadDto(int id, String name, String effect, String description) {
     
 }
