+++ src/main/java/com/poketeambuilder/dtos/front/move/MoveSummaryDto.java
@@ -2,6 +2,6 @@ package com.poketeambuilder.dtos.front.move;
 
 import com.poketeambuilder.dtos.front.type.typing.TypeReadDto;
 
-public record MoveSummaryDto(long id, String name, TypeReadDto type) {
+public record MoveSummaryDto(int id, String name, TypeReadDto type) {
     
 }
