+++ src/main/java/com/poketeambuilder/dtos/front/nature/NatureReadDto.java
@@ -1,5 +1,5 @@
 package com.poketeambuilder.dtos.front.nature;
 
-public record NatureReadDto(long id, String name, String increasedStat, String decreasedStat) {
+public record NatureReadDto(int id, String name, String increasedStat, String decreasedStat) {
     
 }
