+++ src/main/java/com/poketeambuilder/dtos/front/pokemon/individual/PokemonSummaryDto.java
@@ -2,6 +2,6 @@ package com.poketeambuilder.dtos.front.pokemon.individual;
 
 import com.poketeambuilder.dtos.front.type.typing.TypeReadDto;
 
-public record PokemonSummaryDto(long id, String name, Integer order, TypeReadDto primaryType, TypeReadDto secondaryType, String spriteDefault) {
+public record PokemonSummaryDto(int id, String name, Integer order, TypeReadDto primaryType, TypeReadDto secondaryType, String spriteDefault) {
     
 }
