+++ src/main/java/com/poketeambuilder/dtos/front/pokemon/species/PokemonSpeciesSummaryDto.java
@@ -1,5 +1,5 @@
 package com.poketeambuilder.dtos.front.pokemon.species;
 
-public record PokemonSpeciesSummaryDto(long id, String name, Integer nationalDexNumber, Integer order) {
+public record PokemonSpeciesSummaryDto(int id, String name, Integer nationalDexNumber, Integer order) {
     
 }
